Modularize database and deno under data
