+++ payas-server-core/src/graphql/data/deno/deno_resolver.rs
@@ -13,7 +13,7 @@ use payas_model::model::{GqlCompositeType, GqlCompositeTypeKind, GqlTypeKind};
 use crate::graphql::data::access_solver;
 use crate::graphql::data::operation_mapper::DenoOperation;
 
-use crate::graphql::deno::{ClayCallbackProcessor, FnClaytipExecuteQuery};
+use crate::graphql::data::deno::{ClayCallbackProcessor, FnClaytipExecuteQuery};
 use crate::graphql::execution::query_response::{QueryResponse, QueryResponseBody};
 use crate::graphql::execution_error::{ExecutionError, ServiceExecutionError};
 use crate::graphql::request_context::RequestContext;
@@ -46,7 +46,7 @@ impl DenoOperation {
         resolve_deno(
             method,
             field,
-            crate::graphql::deno::claytip_execute_query!(system_context, request_context),
+            super::claytip_execute_query!(system_context, request_context),
             system_context,
             request_context,
         )
