+++ payas-server-core/src/graphql/data/deno/interception.rs
@@ -7,7 +7,7 @@ use payas_deno::Arg;
 use payas_model::model::interceptor::{Interceptor, InterceptorKind};
 
 use crate::graphql::{
-    deno::{
+    data::deno::{
         clay_execution::ClaytipMethodResponse, ClayCallbackProcessor, FnClaytipExecuteQuery,
         FnClaytipInterceptorProceed, InterceptedOperationInfo,
     },
