+++ payas-server-core/src/graphql/data/mod.rs
@@ -1,4 +1,6 @@
 pub mod access_solver;
+pub mod database;
+pub mod deno;
 pub mod mutation_resolver;
 pub mod operation_mapper;
 mod operation_resolver;
