+++ payas-server-core/src/graphql/validation/document_validator.rs
@@ -3,7 +3,9 @@ use async_graphql_value::Name;
 use serde_json::{Map, Value};
 use tracing::instrument;
 
-use crate::graphql::{introspection::schema::Schema, validation_error::ValidationError};
+use crate::graphql::{
+    introspection::schema::Schema, validation::validation_error::ValidationError,
+};
 
 use super::{operation::ValidatedOperation, operation_validator::OperationValidator};
