+++ payas-server-core/src/graphql/validation/selection_set_validator.rs
@@ -14,7 +14,7 @@ use crate::graphql::{
         definition::type_introspection::TypeDefinitionIntrospection,
         schema::{Schema, QUERY_ROOT_TYPENAME},
     },
-    validation_error::ValidationError,
+    validation::validation_error::ValidationError,
 };
 
 use super::{arguments_validator::ArgumentValidator, field::ValidatedField, underlying_type};
